Conflicts:
	rv-predict-agent/src/main/java/com/runtimeverification/rvpredict/instrumentation/Agent.java
	rv-predict-agent/src/main/java/com/runtimeverification/rvpredict/instrumentation/transformer/ClassTransformer.java
	rv-predict-agent/src/main/java/com/runtimeverification/rvpredict/instrumentation/transformer/MethodTransformer.java
	rv-predict-agent/src/main/java/com/runtimeverification/rvpredict/runtime/RVPredictRuntime.java
	rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/log/EventInputStream.java
	rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/log/EventItem.java
	rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/log/OfflineLoggingFactory.java
	rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/log/TraceCache.java
	rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/trace/AbstractEvent.java
	rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/util/DumpLogFile.java

+++ rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/trace/InitEvent.java
@@ -33,7 +33,7 @@ package com.runtimeverification.rvpredict.trace;
  */
 public class InitEvent extends InitOrAccessEvent {
 
-    public InitEvent(long GID, long tid, int ID, long objectHashCode, int index, long value) {
+    public InitEvent(long GID, long tid, int ID, int objectHashCode, int index, long value) {
         super(GID, tid, ID, EventType.INIT, objectHashCode, index, value);
     }
