+++ rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/trace/MemoryAccessEvent.java
@@ -33,7 +33,7 @@ package com.runtimeverification.rvpredict.trace;
  */
 public abstract class MemoryAccessEvent extends InitOrAccessEvent {
 
-    protected MemoryAccessEvent(long GID, long TID, int ID, EventType type, long objectHashCode,
+    protected MemoryAccessEvent(long GID, long TID, int ID, EventType type, int objectHashCode,
             int index, long value) {
         super(GID, TID, ID, type, objectHashCode, index, value);
         assert type == EventType.READ || type == EventType.WRITE;
