+++ rv-predict-engine/src/main/java/com/runtimeverification/rvpredict/trace/WriteEvent.java
@@ -30,7 +30,7 @@ package com.runtimeverification.rvpredict.trace;
 
 public class WriteEvent extends MemoryAccessEvent {
 
-    public WriteEvent(long GID, long tid, int ID, long objectHashCode, int index, long value) {
+    public WriteEvent(long GID, long tid, int ID, int objectHashCode, int index, long value) {
         super(GID, tid, ID, EventType.WRITE, objectHashCode, index, value);
     }
